made logger optional
