+++ examples/react/develop/functional-component-ts/package.json
@@ -5,10 +5,11 @@
   "dependencies": {
     "@agile-ts/api": "file:.yalc/@agile-ts/api",
     "@agile-ts/core": "file:.yalc/@agile-ts/core",
-    "@agile-ts/multieditor": "file:.yalc/@agile-ts/multieditor",
-    "@agile-ts/react": "file:.yalc/@agile-ts/react",
     "@agile-ts/event": "file:.yalc/@agile-ts/event",
+    "@agile-ts/logger": "file:.yalc/@agile-ts/logger",
+    "@agile-ts/multieditor": "file:.yalc/@agile-ts/multieditor",
     "@agile-ts/proxytree": "file:.yalc/@agile-ts/proxytree",
+    "@agile-ts/react": "file:.yalc/@agile-ts/react",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
     "react-router-dom": "^5.2.0",
@@ -30,7 +31,7 @@
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject",
-    "install:agile": "yalc add @agile-ts/core @agile-ts/react @agile-ts/api @agile-ts/multieditor @agile-ts/event @agile-ts/proxytree & yarn install"
+    "install:agile": "yalc add @agile-ts/core @agile-ts/react @agile-ts/api @agile-ts/multieditor @agile-ts/event @agile-ts/proxytree @agile-ts/logger & yarn install"
   },
   "eslintConfig": {
     "extends": "react-app"
