+++ packages/core/src/runtime/index.ts
@@ -71,9 +71,7 @@ export class Runtime {
     // Add specified Job to the queue
     this.jobQueue.push(job);
 
-    Agile.logger.if
-      .tag(['runtime'])
-      .info(LogCodeManager.getLog('16:01:00', [job._key]), job);
+    LogCodeManager.logIfTags(['runtime'], '16:01:00', [job._key], job);
 
     // Run first Job from the queue
     if (config.perform) {
@@ -110,9 +108,7 @@ export class Runtime {
     if (job.rerender) this.jobsToRerender.push(job);
     this.currentJob = null;
 
-    Agile.logger.if
-      .tag(['runtime'])
-      .info(LogCodeManager.getLog('16:01:01', [job._key]), job);
+    LogCodeManager.logIfTags(['runtime'], '16:01:01', [job._key], job);
 
     // Perform Jobs as long as Jobs are left in the queue.
     // If no Job is left start updating (re-rendering) Subscription Container (UI-Components)
@@ -262,9 +258,12 @@ export class Runtime {
       subscriptionContainer.updatedSubscribers.clear();
     });
 
-    Agile.logger.if
-      .tag(['runtime'])
-      .info(LogCodeManager.getLog('16:01:02'), subscriptionsToUpdate);
+    LogCodeManager.logIfTags(
+      ['runtime'],
+      '16:01:02',
+      [],
+      subscriptionsToUpdate
+    );
   }
 
   /**
