+++ packages/core/src/runtime/subscription/sub.controller.ts
@@ -165,9 +165,12 @@ export class SubController {
     if (subscriptionInstance instanceof CallbackSubscriptionContainer) {
       unsub(subscriptionInstance);
       this.callbackSubs.delete(subscriptionInstance);
-      Agile.logger.if
-        .tag(['runtime', 'subscription'])
-        .info(LogCodeManager.getLog('15:01:00'), subscriptionInstance);
+      LogCodeManager.logIfTags(
+        ['subscription'],
+        '15:01:00',
+        [],
+        subscriptionInstance
+      );
       return;
     }
 
@@ -175,9 +178,12 @@ export class SubController {
     if (subscriptionInstance instanceof ComponentSubscriptionContainer) {
       unsub(subscriptionInstance);
       this.componentSubs.delete(subscriptionInstance);
-      Agile.logger.if
-        .tag(['runtime', 'subscription'])
-        .info(LogCodeManager.getLog('15:01:01'), subscriptionInstance);
+      LogCodeManager.logIfTags(
+        ['subscription'],
+        '15:01:01',
+        [],
+        subscriptionInstance
+      );
       return;
     }
 
@@ -191,9 +197,12 @@ export class SubController {
         (subContainer) => {
           unsub(subContainer as ComponentSubscriptionContainer);
           this.componentSubs.delete(subContainer);
-          Agile.logger.if
-            .tag(['runtime', 'subscription'])
-            .info(LogCodeManager.getLog('15:01:01'), subscriptionInstance);
+          LogCodeManager.logIfTags(
+            ['subscription'],
+            '15:01:01',
+            [],
+            subscriptionInstance
+          );
         }
       );
       return;
@@ -241,9 +250,12 @@ export class SubController {
         componentSubscriptionContainer,
       ];
 
-    Agile.logger.if
-      .tag(['runtime', 'subscription'])
-      .info(LogCodeManager.getLog('15:01:02'), componentSubscriptionContainer);
+    LogCodeManager.logIfTags(
+      ['subscription'],
+      '15:01:02',
+      [],
+      componentSubscriptionContainer
+    );
 
     return componentSubscriptionContainer;
   }
@@ -270,9 +282,12 @@ export class SubController {
     this.callbackSubs.add(callbackSubscriptionContainer);
     callbackSubscriptionContainer.ready = true;
 
-    Agile.logger.if
-      .tag(['runtime', 'subscription'])
-      .info(LogCodeManager.getLog('15:01:03'), callbackSubscriptionContainer);
+    LogCodeManager.logIfTags(
+      ['subscription'],
+      '15:01:03',
+      [],
+      callbackSubscriptionContainer
+    );
 
     return callbackSubscriptionContainer;
   }
