+++ packages/core/src/shared.ts
@@ -10,7 +10,6 @@ import {
   defineConfig,
   DependableAgileInstancesType,
   flatMerge,
-  Logger,
   removeProperties,
   runsOnServer,
   State,
@@ -23,7 +22,6 @@ import {
  */
 let sharedAgileInstance = new Agile({
   key: 'shared',
-  logConfig: { prefix: 'Agile', level: Logger.level.ERROR, active: true },
   localStorage: !runsOnServer(),
 });
 export { sharedAgileInstance as shared };
